fix(backend)：移除数据库驱动的运行时站点设置
1. 将 canonical URL、SEO、robots、CORS 和 Waline JWT 迁移到环境变量。
2. 删除运行时站点设置相关 API、模型、种子、测试和迁移。
3. 保持现有页面与部署路径一致，避免影响前端样式。

+++ backend/src/aerisun/domain/activity/schemas.py
@@ -38,7 +38,7 @@ class RecentActivityRead(ModelBase):
 class ActivityHeatmapStatsRead(ModelBase):
     total_contributions: int = Field(description="Total contributions in the period")
     peak_week: int = Field(description="Highest weekly contribution count")
-    average_per_week: int = Field(description="Average contributions per week")
+    average_per_week: float = Field(description="Average contributions per week")
 
 
 class ActivityHeatmapWeekRead(ModelBase):
